Test and fix instruction class
- Only optional thing remaining is argument size checking

+++ lib/instruction_factory.hpp
@@ -30,7 +30,7 @@ void InstructionFactory::register_helper(std::index_sequence<I...>)
             if constexpr (sizeof...(I) > 0)
             {
                 assert(val > begin);
-                return std::make_unique<T>(val - begin);
+                return std::make_unique<T>(static_cast<std::byte>(val - begin));
             }
             else
             {
